# docs/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "智能知识库平台",
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/tasks": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "List Admin Tasks",
        "operationId": "list_admin_tasks_api_v1_admin_tasks_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_PaginatedData_AdminTaskItem__"
                }
              }
            }
          },
          "422": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录、Token 无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "已登录但权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/retrieval/search": {
      "post": {
        "tags": [
          "retrieval"
        ],
        "summary": "Search Endpoint",
        "operationId": "search_endpoint_api_v1_retrieval_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Search Endpoint Api V1 Retrieval Search Post"
                }
              }
            }
          },
          "422": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录、Token 无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "已登录但权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/retrieval/answer": {
      "post": {
        "tags": [
          "retrieval"
        ],
        "summary": "Answer Endpoint",
        "operationId": "answer_endpoint_api_v1_retrieval_answer_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RagAnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Answer Endpoint Api V1 Retrieval Answer Post"
                }
              }
            }
          },
          "422": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录、Token 无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "已登录但权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/notifications": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "List Notifications Endpoint",
        "operationId": "list_notifications_endpoint_api_v1_notifications_get",
        "parameters": [
          {
            "name": "audience",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "user",
                "admin"
              ],
              "type": "string",
              "default": "user",
              "title": "Audience"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Notifications Endpoint Api V1 Notifications Get"
                }
              }
            }
          },
          "422": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录、Token 无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "已登录但权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/notifications/{notification_id}/read": {
      "patch": {
        "tags": [
          "notifications"
        ],
        "summary": "Mark Notification Read Endpoint",
        "operationId": "mark_notification_read_endpoint_api_v1_notifications__notification_id__read_patch",
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Mark Notification Read Endpoint Api V1 Notifications  Notification Id  Read Patch"
                }
              }
            }
          },
          "422": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录、Token 无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "已登录但权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/notifications/mark-all-read": {
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Mark All Notifications Read Endpoint",
        "operationId": "mark_all_notifications_read_endpoint_api_v1_notifications_mark_all_read_post",
        "parameters": [
          {
            "name": "audience",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "user",
                "admin"
              ],
              "type": "string",
              "default": "user",
              "title": "Audience"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Mark All Notifications Read Endpoint Api V1 Notifications Mark All Read Post"
                }
              }
            }
          },
          "422": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录、Token 无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "已登录但权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "APIResponse_PaginatedData_AdminTaskItem__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginatedData_AdminTaskItem_"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "APIResponse[PaginatedData[AdminTaskItem]]"
      },
...
      "AdminTaskItem": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "task_type": {
            "type": "string",
            "title": "Task Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "retry_count": {
            "type": "integer",
            "title": "Retry Count"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "document_title": {
            "type": "string",
            "title": "Document Title"
          },
          "knowledge_base_id": {
            "type": "string",
            "title": "Knowledge Base Id"
          },
          "knowledge_base_name": {
            "type": "string",
            "title": "Knowledge Base Name"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "task_type",
          "status",
          "stage",
          "progress",
          "retry_count",
          "request_id",
          "document_id",
          "document_title",
          "knowledge_base_id",
          "knowledge_base_name"
        ],
        "title": "AdminTaskItem"
      },
...
      "PaginatedData_AdminTaskItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminTaskItem"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "PaginatedData[AdminTaskItem]"
      },
...
      "RagAnswerRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query"
          },
          "mode": {
            "type": "string",
            "enum": [
              "keyword",
              "vector",
              "hybrid"
            ],
            "title": "Mode"
          },
          "kb_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kb Id"
          },
          "top_k": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 10
          },
          "threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Threshold",
            "default": 0.0
          },
          "metadata_filter": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Filter"
          },
          "rerank": {
            "type": "boolean",
            "title": "Rerank",
            "default": true
          },
          "rerank_model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rerank Model Id"
          },
          "embedding_model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Model Id"
          },
          "chat_model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Model Id"
          }
        },
        "type": "object",
        "required": [
          "query",
          "mode"
        ],
        "title": "RagAnswerRequest"
      },
...
      "SearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query"
          },
          "mode": {
            "type": "string",
            "enum": [
              "keyword",
              "vector",
              "hybrid"
            ],
            "title": "Mode"
          },
          "kb_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kb Id"
          },
          "top_k": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 10
          },
          "threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Threshold",
            "default": 0.0
          },
          "metadata_filter": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Filter"
          },
          "rerank": {
            "type": "boolean",
            "title": "Rerank",
            "default": true
          },
          "rerank_model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rerank Model Id"
          },
          "embedding_model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Model Id"
          }
        },
        "type": "object",
        "required": [
          "query",
          "mode"
        ],
        "title": "SearchRequest"
      },
...
      "APIErrorResponse": {
        "type": "object",
        "title": "APIErrorResponse",
        "description": "所有非健康检查 API 的统一错误响应。",
        "required": [
          "code",
          "message",
          "data",
          "request_id"
        ],
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "null"
          },
          "request_id": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API 前缀"
    }
  ]
}